* More anx mux fixes...
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8495 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/core/ogg/libOOOgg/StampedOggPacket.cpp
@@ -120,5 +120,5 @@ void StampedOggPacket::setEndTime(LOOG_INT64 inEndTime) {
 void StampedOggPacket::setTimeStamp(LOOG_INT64 inStartTime, LOOG_INT64 inEndTime, StampedOggPacket::eStampType inStampType) {
 	mStartTime = inStartTime;
 	mEndTime = inEndTime;
-	mStampType = inStampType;
+	mStampType = (unsigned short)inStampType;
 }
